requirements.txt
xmltodict 
vosk 
pydub
langchain-community
langchain-core
langchain_huggingface
faiss-cpu
google-generativeai
python-dotenv
streamlit
langchain_google_genai